Consolidate shared logic into agent-setup.ts
