+++ src/index.ts
@@ -3,7 +3,6 @@ import type {
   ExtensionContext,
 } from "@mariozechner/pi-coding-agent";
 import path from "node:path";
-import { Type } from "@sinclair/typebox";
 import { parseArgs } from "citty";
 import {
   claimEvent,
@@ -25,6 +24,7 @@ import {
   buildAgentSystemPrompt,
   registerAgentMemoryTool,
   registerAgentHooks,
+  registerBusytownTools,
   resolveAgentModel,
 } from "./agent-setup.ts";
 import { listAgentDefs, loadAgentDef } from "./file-agent.ts";
@@ -79,95 +79,7 @@ export default (pi: ExtensionAPI) => {
     registerEventLogCommand(pi, db);
 
     // Register tools
-    pi.registerTool({
-      name: "busytown-push",
-      label: "Busytown Push",
-      description:
-        "Push an event to the Busytown event queue to trigger agent workflows. " +
-        "Common event types: plan.request, code.request, review.request",
-      parameters: Type.Object({
-        type: Type.String({ description: "Event type (e.g. 'plan.request')" }),
-        payload: Type.Optional(
-          Type.String({ description: "JSON payload string (default: '{}')" }),
-        ),
-      }),
-      execute: async (_toolCallId, params) => {
-        await nextTick();
-        const payload = params.payload
-          ? JSON.parse(params.payload as string)
-          : {};
-        const event = pushEvent(db, "pi", params.type as string, payload);
-        return {
-          content: [{ type: "text", text: JSON.stringify(event, null, 2) }],
-          details: {},
-        };
-      },
-    });
-
-    pi.registerTool({
-      name: "busytown-events",
-      label: "Busytown Events",
-      description: "List recent events from the Busytown event queue",
-      parameters: Type.Object({
-        type: Type.Optional(
-          Type.String({ description: "Filter by event type" }),
-        ),
-        tail: Type.Optional(
-          Type.Integer({
-            description: "Number of recent events to show (default: 20)",
-          }),
-        ),
-        since: Type.Optional(
-          Type.Integer({
-            description:
-              "List events since this event ID (mutually exclusive with tail)",
-          }),
-        ),
-      }),
-      async execute(_toolCallId, params) {
-        await nextTick();
-        const sinceId = params.since as number | undefined;
-        const events = getEventsSince(db, {
-          ...(sinceId != null
-            ? { sinceId }
-            : { tail: (params.tail as number) ?? 20 }),
-          filterType: params.type as string,
-        });
-        const ljson = events.map((e) => JSON.stringify(e)).join("\n");
-        return {
-          content: [{ type: "text", text: ljson }],
-          details: {},
-        };
-      },
-    });
-
-    pi.registerTool({
-      name: "busytown-claim",
-      label: "Busytown Claim",
-      description: "Claim an event so no other agent processes it",
-      parameters: Type.Object({
-        event_id: Type.Integer({ description: "Event ID to claim" }),
-        agent: Type.String({ description: "Agent ID claiming the event" }),
-      }),
-      async execute(_toolCallId, params) {
-        await nextTick();
-        const claimed = claimEvent(
-          db,
-          params.agent as string,
-          params.event_id as number,
-        );
-        const claimant = getClaimant(db, params.event_id as number);
-        return {
-          content: [
-            {
-              type: "text",
-              text: JSON.stringify({ claimed, claimant }, null, 2),
-            },
-          ],
-          details: {},
-        };
-      },
-    });
+    registerBusytownTools(pi, db, "pi");
 
     // Register commands (slash-command equivalents of the tools)
